fix: use latest playwright image (#612)
The default playwright image hasn't been updated for months so we end up removing and reinstalling browsers on every run, so use a specific version instead.

The playwright image comes with LTS node so we can base both the node and browser images off that and save having to build the images over and over again.